SMIL-55 | SMIL-58 | SMIL-68: Resume upload & download fixes

+++ src/main/java/com/smilebat/learntribe/processor/services/experiences/DefaultExperienceStrategy.java
@@ -3,6 +3,7 @@ package com.smilebat.learntribe.processor.services.experiences;
 import com.smilebat.learntribe.dataaccess.jpa.entity.Experience;
 import java.util.Collection;
 import java.util.Set;
+import java.util.function.Consumer;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
@@ -18,15 +19,22 @@ import lombok.extern.slf4j.Slf4j;
 public class DefaultExperienceStrategy<P extends ExperienceContext, R extends Experience>
     implements ExperienceStartegy<P> {
 
+  private Consumer<Experience> resetId = (exp) -> exp.setId(null);
+
   @Override
   public void updateExperiences(P context) {
     log.info("Updating User Experiences");
     Collection<R> existingExperiences = context.getExistingExperiences();
     Collection<R> updatedExperiences = context.getRequestExperiences();
-    context.setUpdatedExperiences(updatedExperiences);
     Set<Long> deletedExperienceIds =
         evaluateDeletedExperienceIds(existingExperiences, updatedExperiences);
-    context.getRepository().deleteAllById(deletedExperienceIds);
+    if (!deletedExperienceIds.isEmpty()) {
+      context.getRepository().deleteAllById(deletedExperienceIds);
+    }
+    if (deletedExperienceIds.isEmpty() && existingExperiences.isEmpty()) {
+      updatedExperiences.forEach(resetId);
+    }
+    context.setUpdatedExperiences(updatedExperiences);
   }
 
   protected Set<Long> evaluateDeletedExperienceIds(
